[Feat] 홈뷰 그룹 변경시 초성확인하기 버튼 비활성화, SideBarView 구성원 관리 리스트 접고 열때 애니메이션 추가, CommMainView 구성원리스트 검색, 취소시 깜빡거리는 문제 해결, ZenoBackBtn 전체가 클릭되는 문제 해결, SettingTextFieldView 글자변경 없을 시 disable처리, ShareText함수 ViewModel로 이동, HomeMainView 뷰 분리
