+++ Zeno/Sources/View/Alarm/AlarmView.swift
@@ -159,6 +159,9 @@ struct AlarmView: View {
                 }
             }
         }
+        .onChange(of: selectedCommunityId) { _ in
+            selectAlarm = nil
+        }
     }
 }
