+++ Zeno/Sources/View/Community/HamburgerBar/CommUserMgmtView.swift
@@ -12,59 +12,59 @@ struct CommUserMgmtView: View {
     @Environment(\.dismiss) private var dismiss
     @EnvironmentObject private var commViewModel: CommViewModel
     
-    @State private var isDeportAlert = false
     @State private var deportUser = User.emptyUser
+    @State private var isDeportAlert = false
     @State private var isWaitListFold = false
     @State private var isCurrentListFold = false
     
     var body: some View {
-        VStack {
-            ZenoNavigationBackBtn {
-                dismiss()
-            } label: {
-                HStack {
-                    Text("구성원 관리")
-                    Spacer()
+            VStack {
+                ZenoNavigationBackBtn {
+                    dismiss()
+                } tailingLabel: {
+                    HStack {
+                        Text("구성원 관리")
+                        Spacer()
+                    }
                 }
-            }
-            ScrollView(showsIndicators: false) {
-                ForEach(MGMTSection.allCases) { section in
-                    switch section {
-                    case .wait:
-                        ZenoProfileVisibleListView(list: commViewModel.currentWaitApprovalMembers) {
-                            Text("\(section.header) \(commViewModel.currentWaitApprovalMembers.count)")
-                        } btnLabel: {
-                            HStack(alignment: .bottom, spacing: 2) {
-                                Image(systemName: "person.crop.circle.badge.plus")
-                                Text("가입수락")
+                ScrollView(showsIndicators: false) {
+                    ForEach(MGMTSection.allCases) { section in
+                        switch section {
+                        case .wait:
+                            ZenoProfileFoldableListView(isListFold: $isWaitListFold, list: commViewModel.currentWaitApprovalMembers) {
+                                Text("\(section.header) \(commViewModel.currentWaitApprovalMembers.count)")
+                            } btnLabel: {
+                                HStack(alignment: .bottom, spacing: 2) {
+                                    Image(systemName: "person.crop.circle.badge.plus")
+                                    Text("가입수락")
+                                }
+                            } interaction: { user in
+                                Task {
+                                    await commViewModel.acceptMember(user: user)
+                                }
                             }
-                        } interaction: { user in
-                            Task {
-                                await commViewModel.acceptMember(user: user)
+                        case .general:
+                            ZenoProfileFoldableListView(isListFold: $isCurrentListFold, list: commViewModel.currentCommMembers) {
+                                Text("\(section.header) \(commViewModel.currentCommMembers.count)")
+                            } btnLabel: {
+                                HStack(alignment: .bottom, spacing: 2) {
+                                    Image(systemName: "person.crop.circle.badge.minus")
+                                    Text("추방하기")
+                                }
+                            } interaction: { user in
+                                deportUser = user
+                                isDeportAlert = true
                             }
-                        }
-                    case .general:
-                        ZenoProfileVisibleListView(list: commViewModel.currentCommMembers) {
-                            Text("\(section.header) \(commViewModel.currentCommMembers.count)")
-                        } btnLabel: {
-                            HStack(alignment: .bottom, spacing: 2) {
-                                Image(systemName: "person.crop.circle.badge.minus")
-                                Text("추방하기")
-                            }
-                        } interaction: { user in
-                            deportUser = user
-                            isDeportAlert = true
                         }
                     }
                 }
-            }
-            .refreshable {
-                Task {
-                    await commViewModel.fetchCurrentCommMembers()
-                }
+                .animation(.easeInOut, value: [isWaitListFold, isCurrentListFold])
+                .refreshable {
+                    Task {
+                        await commViewModel.fetchCurrentCommMembers()
+                    }
             }
         }
-        .tint(.black)
         .alert("\(deportUser.name)님을 내보낼까요?", isPresented: $isDeportAlert) {
             Button("내보내기", role: .destructive) {
                 Task {
